Update core - Refactor CPU processor.
