+++ core/src/processor/cpu/Generic.cpp
@@ -1,5 +1,6 @@
+#include <cstdint>
+
 #include "AC/Core/Image.hpp"
-#include "AC/Core/Util.hpp"
 
 #include "AC/Core/Internal/Processor/CPU/Common.hpp"
