+++ core/src/processor/cpu/arm/NEON.cpp
@@ -1,7 +1,8 @@
+#include <cstdint>
+
 #include <arm_neon.h>
 
 #include "AC/Core/Image.hpp"
-#include "AC/Core/Util.hpp"
 #include "AC/Util/Macro.hpp"
 
 #include "AC/Core/Internal/Processor/CPU/Common.hpp"
