+++ core/src/processor/cpu/wasm/SIMD128.cpp
@@ -1,7 +1,8 @@
+#include <cstdint>
+
 #include <wasm_simd128.h>
 
 #include "AC/Core/Image.hpp"
-#include "AC/Core/Util.hpp"
 #include "AC/Util/Macro.hpp"
 
 #include "AC/Core/Internal/Processor/CPU/Common.hpp"
