+++ core/src/processor/cpu/x86/AVX.cpp
@@ -1,8 +1,9 @@
+#include <cstdint>
+
 #include <immintrin.h>
 
 #include "AC/Core/SIMD.hpp"
 #include "AC/Core/Image.hpp"
-#include "AC/Core/Util.hpp"
 #include "AC/Util/Macro.hpp"
 
 #include "AC/Core/Internal/Processor/CPU/Common.hpp"
